rearrange manager application to allow for multiple storage backends
